+++ pylinac/core/geometry.py
@@ -186,7 +186,12 @@ class Circle:
         return self.radius * 2
 
     def plot2axes(
-        self, axes: plt.Axes, edgecolor: str = "black", fill: bool = False
+        self,
+        axes: plt.Axes,
+        edgecolor: str = "black",
+        fill: bool = False,
+        text: str = "",
+        fontsize: str = "medium",
     ) -> None:
         """Plot the Circle on the axes.
 
@@ -198,6 +203,11 @@ class Circle:
             The color of the circle.
         fill : bool
             Whether to fill the circle with color or leave hollow.
+        text: str
+            If provided, plots the given text at the center. Useful for identifying ROIs on a plotted image apart.
+        fontsize: str
+            The size of the text, if provided. See https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.text.html
+            for options.
         """
         axes.add_patch(
             mpl_Circle(
@@ -207,6 +217,22 @@ class Circle:
                 fill=fill,
             )
         )
+        if text:
+            axes.text(
+                x=self.center.x,
+                y=self.center.y,
+                s=text,
+                fontsize=fontsize,
+                color=edgecolor,
+            )
+
+    def as_dict(self) -> dict:
+        """Convert to dict. Useful for dataclasses/Result"""
+        return {
+            "center_x": self.center.x,
+            "center_y": self.center.y,
+            "diameter": self.diameter,
+        }
 
 
 class Vector: